auth-service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth-oauthlib==0.4.6
python-dotenv==0.19.0
requests==2.26.0
